+++ pkg/template_binding/CHANGELOG.md
@@ -3,15 +3,20 @@
 This file contains highlights of what changes on each version of the
 template_binding package.
 
+#### Pub version 0.12.0-pre.1.dev
+  * Fix bug in translation of text to textContent (in case custom-elements use
+    `text` as an attribute name).
+
 #### Pub version 0.12.0-dev
   * NodeBind interop support. This allows elements such as Polymer's
     core-elements and paper-elements to work properly with Dart binding paths,
     including using Elements and functions as values, and two-way bindings.
   * NodeBind is no longer ported. It now comes from
     packages/web_components/platform.js
+  * Up to date with [TemplateBinding#d9f4543][d9f4543] (release 0.3.4)
 
 #### Pub version 0.11.0
-  * Ported up to commit [TemplateBinding#1cee02][5b9a3b] and
+  * Ported up to commit [TemplateBinding#5b9a3b][5b9a3b] and
     [NodeBind#c47bc1][c47bc1].
 
 #### Pub version 0.10.0
@@ -22,7 +27,8 @@ template_binding package.
   * Ported up to commit [TemplateBinding#99e52d][99e52d] and
     [NodeBind#f7cc76][f7cc76].
 
-[1cee02]: https://github.com/Polymer/TemplateBinding/commit/5b9a3be40682e1ccd5e6c0b04fbe2c54d74b5d1e
+[d9f4543]: https://github.com/Polymer/TemplateBinding/commit/d9f4543dc06935824bfd43564c442b0897ce1c54
+[5b9a3b]: https://github.com/Polymer/TemplateBinding/commit/5b9a3be40682e1ccd5e6c0b04fbe2c54d74b5d1e
 [c47bc1]: https://github.com/Polymer/NodeBind/commit/c47bc1b40d1cf0123b29620820a7111471e83ff3
 [51df59]: https://github.com/Polymer/TemplateBinding/commit/51df59c16e0922dec041cfe604016aac00918d5d
 [99e52d]: https://github.com/Polymer/TemplateBinding/commit/99e52dd7fbaefdaee9807648d1d6097eb3e99eda